feat(optimize): ship structure optimize on the shared compute worker (optimize-worker-ship)
Single execution path: page/vsc-ext -> runOptimize -> optimize/worker_client -> core WorkloadHost. Writeback clones every atom column and re-attaches the cell (no more box/column loss), publishes via an explicit DataSource.replaceHeadFrame seam, and rejects triclinic cells instead of squaring them. Worker builds Box.ortho from all three edges; typified frames keep the cell so PBC reaches the force field. Cancel posts exactly once and resolves with last coordinates; dispose during boot rejects pending runs. Main-thread relax entries pruned from the stage root. Typecheck gate now covers tests in core and stage (75 pre-existing test type errors fixed).

+++ stage/tests/artist/bond_buffer.test.ts
@@ -96,7 +96,7 @@ function makeBenzene(): { atoms: Block; bonds: Block; normal: Vector3 } {
 }
 
 function makeAtomColor(count: number): Float32Array {
-  const color = new Float64Array(count * 4);
+  const color = new Float32Array(count * 4);
   for (let i = 0; i < count; i++) {
     color[i * 4 + 0] = 0.5;
     color[i * 4 + 1] = 0.5;
